refactor(safari): download dnr validator on demand, drop postinstall hook
