+++ package.json
@@ -4,7 +4,6 @@
   "version": "10.5.39",
   "type": "module",
   "scripts": {
-    "postinstall": "node scripts/download-validate-dnr-rules.js",
     "build": "node scripts/build.js",
     "start": "npm run build -- --watch",
     "start:update": "./scripts/update.sh",
